refactor: move create group to creation directory
